Update convert to use additive notation
A simpler form of roman numerals is additive notation where the unit 
value of 4 employs the symbol 'IIII' instead of 'IV'.

Use a unit list with values in descending order to loop through, check 
if the number arguement is larger than the unit, add the numeral symbol 
to a string and subtract the unit from the number. Repeat this until 
number is 0.