Update convert to use subtractive notation
The subtractive notation uses abbreviated symbols for a number that is 1 
unit value less than the unit value that is one step larger.

Add a check to the loop for this condition that then creates the correct 
symbol for the number. The units list will need to be enumerated so that 
the index can be accessed.